fix(lua): move bindings to cpp files
