+++ src/lua/bindings/vector.hpp
@@ -29,14 +29,5 @@ namespace lua::vector
 	// Field: z: float
 	// z component of the vector.
 
-	static void bind(sol::state& state)
-	{
-		//clang-format off
-		state.new_usertype<Vector3>("vec3",
-			sol::constructors<Vector3(float, float, float)>(),
-			"x", &Vector3::x, "y", &Vector3::y, "z", &Vector3::z,
-			"__tostring", &Vector3::to_string
-		);
-		//clang-format on
-	}
+	void bind(sol::state& state);
 }
